+++ src/mroute.h
@@ -111,14 +111,14 @@ extern int mroute4_socket;
  */
 extern int mroute6_socket;
 
-int  mroute4_enable    (int do_vifs);
+int  mroute4_enable    (int do_vifs, int table_id);
 void mroute4_disable   (void);
 int  mroute4_dyn_add   (struct mroute4 *mroute);
 void mroute4_dyn_expire(int max_idle);
 int  mroute4_add       (struct mroute4 *mroute);
 int  mroute4_del       (struct mroute4 *mroute);
 
-int  mroute6_enable    (int do_vifs);
+int  mroute6_enable    (int do_vifs, int table_id);
 void mroute6_disable   (void);
 int  mroute6_add       (struct mroute6 *mroute);
 int  mroute6_del       (struct mroute6 *mroute);